Remove the PathEffect::MakeStroke() method and add a applyToPath() method to Stroke.
