+++ src/core/TrimPathEffect.h
@@ -1,0 +1,35 @@
+/////////////////////////////////////////////////////////////////////////////////////////////////
+//
+//  Tencent is pleased to support the open source community by making tgfx available.
+//
+//  Copyright (C) 2024 THL A29 Limited, a Tencent company. All rights reserved.
+//
+//  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
+//  in compliance with the License. You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/BSD-3-Clause
+//
+//  unless required by applicable law or agreed to in writing, software distributed under the
+//  license is distributed on an "as is" basis, without warranties or conditions of any kind,
+//  either express or implied. see the license for the specific language governing permissions
+//  and limitations under the license.
+//
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include "tgfx/core/PathEffect.h"
+
+namespace tgfx {
+class TrimPathEffect : public PathEffect {
+ public:
+  TrimPathEffect(float startT, float stopT) : startT(startT), stopT(stopT) {
+  }
+
+  bool filterPath(Path* path) const override;
+
+ private:
+  float startT = 0.0f;
+  float stopT = 1.0f;
+};
+}  // namespace tgfx
